requirements.txt
Package                Version
---------------------- ------------
asttokens              3.0.0
attrs                  25.4.0
by                     0.0.7
certifi                2025.8.3
cffi                   2.0.0
charset-normalizer     3.4.3
colorama               0.4.6
comm                   0.2.3
cryptography           43.0.3
debugpy                1.8.17
decorator              5.2.1
exceptiongroup         1.3.0
executing              2.2.1
filelock               3.19.1
fsspec                 2025.9.0
greenlet               3.2.4
h11                    0.16.0
huggingface-hub        0.35.0
idna                   3.10
importlib_metadata     8.7.0
ipykernel              6.31.0
ipython                8.18.1
jedi                   0.19.2
Jinja2                 3.1.4
joblib                 1.5.2
jupyter_client         8.6.3
jupyter_core           5.8.1
keyboard               0.13.5
MarkupSafe             2.1.5
matplotlib-inline      0.1.7
MouseInfo              0.1.3
mpmath                 1.3.0
mysql-connector        2.2.9
mysql-connector-python 9.4.0
nest-asyncio           1.6.0
networkx               3.2.1
numpy                  1.26.4
outcome                1.3.0.post0
packaging              25.0
pandas                 2.3.2
parso                  0.8.5
pillow                 11.2.1
pip                    25.2
platformdirs           4.4.0
prompt_toolkit         3.0.52
protobuf               3.20.3
psutil                 7.1.1
pure_eval              0.2.3
PyAutoGUI              0.9.54
pycparser              2.23
PyGetWindow            0.0.9
Pygments               2.19.2
PyMsgBox               1.0.9
PyMySQL                1.1.2
pyperclip              1.9.0
PyRect                 0.2.0
PyScreeze              1.0.1
PySocks                1.7.1
python-dateutil        2.9.0.post0
python-dotenv          1.1.1
pytweening             1.2.0
pytz                   2025.2
pywin32                311
PyYAML                 6.0.2
pyzmq                  27.1.0
regex                  2025.9.18
requests               2.32.5
safetensors            0.6.2
scikit-learn           1.6.1
scipy                  1.13.1
selenium               4.36.0
sentencepiece          0.1.99
setuptools             49.2.1
six                    1.17.0
sniffio                1.3.1
sortedcontainers       2.4.0
SQLAlchemy             2.0.43
stack-data             0.6.3
sympy                  1.13.3
threadpoolctl          3.6.0
tokenizers             0.22.1
torch                  2.7.1+cu118
torchaudio             2.7.1+cu118
torchvision            0.22.1+cu118
tornado                6.5.2
tqdm                   4.67.1
traitlets              5.14.3
transformers           4.56.2
trio                   0.31.0
trio-websocket         0.12.2
typing_extensions      4.15.0
tzdata                 2025.2
unicode                2.9
Unidecode              1.4.0
urllib3                2.5.0
wcwidth                0.2.14
webdriver-manager      4.0.2
websocket-client       1.9.0
wsproto                1.2.0
zipp                   3.23.0
